#94 fix unrestricted assert and split out go to testcases
